refactor error locations
