+++ src/handler/folder_handler.rs
@@ -1,13 +1,13 @@
 use crate::guard::{Auth, ValidateResult};
+use crate::model::error::folder_errors::{
+    CreateFolderError, DeleteFolderError, GetFolderError, UpdateFolderError,
+};
 use crate::model::request::folder_requests::{CreateFolderRequest, UpdateFolderRequest};
 use crate::model::response::folder_responses::{
     CreateFolderResponse, DeleteFolderResponse, GetFolderResponse, UpdateFolderResponse,
 };
 use crate::model::response::BasicMessage;
 use crate::service::folder_service;
-use crate::service::folder_service::{
-    CreateFolderError, DeleteFolderError, GetFolderError, UpdateFolderError,
-};
 use rocket::serde::json::Json;
 
 #[get("/<id>")]
