+++ src/model/error/guard_errors.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub enum AuthError {
+    Missing,
+    Invalid,
+}
