+++ src/model/error/metadata_errors.rs
@@ -1,0 +1,5 @@
+#[derive(PartialEq)]
+pub enum CreatePasswordError {
+    AlreadyExists,
+    Failure,
+}
