+++ src/model/error/mod.rs
@@ -1,0 +1,4 @@
+pub mod file_errors;
+pub mod folder_errors;
+pub mod guard_errors;
+pub mod metadata_errors;
