+++ src/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod db;
+pub mod error;
 pub mod request;
 pub mod response;
